lets put fvn in fixed raste calculation
