+++ modules/product/src/main/java/com/opengamma/strata/product/swap/FixedRateCalculation.java
@@ -33,6 +33,7 @@ import com.opengamma.strata.basics.index.Index;
 import com.opengamma.strata.basics.schedule.Schedule;
 import com.opengamma.strata.basics.schedule.SchedulePeriod;
 import com.opengamma.strata.basics.value.ValueSchedule;
+import com.opengamma.strata.collect.ArgChecker;
 import com.opengamma.strata.collect.array.DoubleArray;
 import com.opengamma.strata.product.rate.FixedRateComputation;
 import com.opengamma.strata.product.rate.RateComputation;
@@ -83,6 +84,14 @@ public final class FixedRateCalculation
    */
   @PropertyDefinition(get = "optional")
   private final FixedRateStubCalculation finalStub;
+  /**
+   * The future value notional.
+   * <p>
+   * This property is used when the fixed leg of a swap has a future value notional.
+   * This is typically used for Brazilian swaps.
+   */
+  @PropertyDefinition(get = "optional")
+  private final FutureValueNotional futureValueNotional;
 
   //-------------------------------------------------------------------------
   /**
@@ -133,6 +142,16 @@ public final class FixedRateCalculation
     DoubleArray resolvedRates = rate.resolveValues(accrualSchedule);
     // build accrual periods
     ImmutableList.Builder<RateAccrualPeriod> accrualPeriods = ImmutableList.builder();
+    // is future value notional present
+    if (getFutureValueNotional().isPresent()) {
+      ArgChecker.isTrue(accrualSchedule.size() == 1, "only one accrual period when future value notional present");
+      SchedulePeriod period = accrualSchedule.getPeriod(0);
+      double yearFraction = period.yearFraction(dayCount, accrualSchedule);
+      RateComputation rateComputation = FixedOvernightCompoundedAnnualRateComputation.of(rate, yearFraction);
+      accrualPeriods.add(
+          RateAccrualPeriod.builder(period).yearFraction(yearFraction).rateComputation(rateComputation).build());
+      return accrualPeriods.build();
+    }
     for (int i = 0; i < accrualSchedule.size(); i++) {
       SchedulePeriod period = accrualSchedule.getPeriod(i);
       double yearFraction = period.yearFraction(dayCount, accrualSchedule);
@@ -180,13 +199,15 @@ public final class FixedRateCalculation
       DayCount dayCount,
       ValueSchedule rate,
       FixedRateStubCalculation initialStub,
-      FixedRateStubCalculation finalStub) {
+      FixedRateStubCalculation finalStub,
+      FutureValueNotional futureValueNotional) {
     JodaBeanUtils.notNull(dayCount, "dayCount");
     JodaBeanUtils.notNull(rate, "rate");
     this.dayCount = dayCount;
     this.rate = rate;
     this.initialStub = initialStub;
     this.finalStub = finalStub;
+    this.futureValueNotional = futureValueNotional;
   }
 
   @Override
@@ -249,6 +270,20 @@ public final class FixedRateCalculation
 
   //-----------------------------------------------------------------------
   /**
+   * Gets the future value notional.
+   * <p>
+   * The final stub of a swap may have a different rate from the regular accrual periods.
+   * This property allows the stub rate to be specified, either as a known amount or a rate.
+   * If this property is not present, then the rate derived from the {@code rate} property applies during the stub.
+   * If this property is present and there is no initial stub, it is ignored.
+   * @return the optional value of the property, not null
+   */
+  public Optional<FutureValueNotional> getFutureValueNotional() {
+    return Optional.ofNullable(futureValueNotional);
+  }
+
+  //-----------------------------------------------------------------------
+  /**
    * Returns a builder that allows this bean to be mutated.
    * @return the mutable builder, not null
    */
@@ -266,7 +301,8 @@ public final class FixedRateCalculation
       return JodaBeanUtils.equal(dayCount, other.dayCount) &&
           JodaBeanUtils.equal(rate, other.rate) &&
           JodaBeanUtils.equal(initialStub, other.initialStub) &&
-          JodaBeanUtils.equal(finalStub, other.finalStub);
+          JodaBeanUtils.equal(finalStub, other.finalStub) &&
+          JodaBeanUtils.equal(futureValueNotional, other.futureValueNotional);
     }
     return false;
   }
@@ -278,17 +314,19 @@ public final class FixedRateCalculation
     hash = hash * 31 + JodaBeanUtils.hashCode(rate);
     hash = hash * 31 + JodaBeanUtils.hashCode(initialStub);
     hash = hash * 31 + JodaBeanUtils.hashCode(finalStub);
+    hash = hash * 31 + JodaBeanUtils.hashCode(futureValueNotional);
     return hash;
   }
 
   @Override
   public String toString() {
-    StringBuilder buf = new StringBuilder(160);
+    StringBuilder buf = new StringBuilder(192);
     buf.append("FixedRateCalculation{");
     buf.append("dayCount").append('=').append(dayCount).append(',').append(' ');
     buf.append("rate").append('=').append(rate).append(',').append(' ');
     buf.append("initialStub").append('=').append(initialStub).append(',').append(' ');
-    buf.append("finalStub").append('=').append(JodaBeanUtils.toString(finalStub));
+    buf.append("finalStub").append('=').append(finalStub).append(',').append(' ');
+    buf.append("futureValueNotional").append('=').append(JodaBeanUtils.toString(futureValueNotional));
     buf.append('}');
     return buf.toString();
   }
@@ -324,6 +362,11 @@ public final class FixedRateCalculation
     private final MetaProperty<FixedRateStubCalculation> finalStub = DirectMetaProperty.ofImmutable(
         this, "finalStub", FixedRateCalculation.class, FixedRateStubCalculation.class);
     /**
+     * The meta-property for the {@code futureValueNotional} property.
+     */
+    private final MetaProperty<FutureValueNotional> futureValueNotional = DirectMetaProperty.ofImmutable(
+        this, "futureValueNotional", FixedRateCalculation.class, FutureValueNotional.class);
+    /**
      * The meta-properties.
      */
     private final Map<String, MetaProperty<?>> metaPropertyMap$ = new DirectMetaPropertyMap(
@@ -331,7 +374,8 @@ public final class FixedRateCalculation
         "dayCount",
         "rate",
         "initialStub",
-        "finalStub");
+        "finalStub",
+        "futureValueNotional");
 
     /**
      * Restricted constructor.
@@ -350,6 +394,8 @@ public final class FixedRateCalculation
           return initialStub;
         case 355242820:  // finalStub
           return finalStub;
+        case -282775858:  // futureValueNotional
+          return futureValueNotional;
       }
       return super.metaPropertyGet(propertyName);
     }
@@ -402,6 +448,14 @@ public final class FixedRateCalculation
       return finalStub;
     }
 
+    /**
+     * The meta-property for the {@code futureValueNotional} property.
+     * @return the meta-property, not null
+     */
+    public MetaProperty<FutureValueNotional> futureValueNotional() {
+      return futureValueNotional;
+    }
+
     //-----------------------------------------------------------------------
     @Override
     protected Object propertyGet(Bean bean, String propertyName, boolean quiet) {
@@ -414,6 +468,8 @@ public final class FixedRateCalculation
           return ((FixedRateCalculation) bean).initialStub;
         case 355242820:  // finalStub
           return ((FixedRateCalculation) bean).finalStub;
+        case -282775858:  // futureValueNotional
+          return ((FixedRateCalculation) bean).futureValueNotional;
       }
       return super.propertyGet(bean, propertyName, quiet);
     }
@@ -439,6 +495,7 @@ public final class FixedRateCalculation
     private ValueSchedule rate;
     private FixedRateStubCalculation initialStub;
     private FixedRateStubCalculation finalStub;
+    private FutureValueNotional futureValueNotional;
 
     /**
      * Restricted constructor.
@@ -455,6 +512,7 @@ public final class FixedRateCalculation
       this.rate = beanToCopy.getRate();
       this.initialStub = beanToCopy.initialStub;
       this.finalStub = beanToCopy.finalStub;
+      this.futureValueNotional = beanToCopy.futureValueNotional;
     }
 
     //-----------------------------------------------------------------------
@@ -469,6 +527,8 @@ public final class FixedRateCalculation
           return initialStub;
         case 355242820:  // finalStub
           return finalStub;
+        case -282775858:  // futureValueNotional
+          return futureValueNotional;
         default:
           throw new NoSuchElementException("Unknown property: " + propertyName);
       }
@@ -489,6 +549,9 @@ public final class FixedRateCalculation
         case 355242820:  // finalStub
           this.finalStub = (FixedRateStubCalculation) newValue;
           break;
+        case -282775858:  // futureValueNotional
+          this.futureValueNotional = (FutureValueNotional) newValue;
+          break;
         default:
           throw new NoSuchElementException("Unknown property: " + propertyName);
       }
@@ -507,7 +570,8 @@ public final class FixedRateCalculation
           dayCount,
           rate,
           initialStub,
-          finalStub);
+          finalStub,
+          futureValueNotional);
     }
 
     //-----------------------------------------------------------------------
@@ -569,15 +633,31 @@ public final class FixedRateCalculation
       return this;
     }
 
+    /**
+     * Sets the future value notional.
+     * <p>
+     * The final stub of a swap may have a different rate from the regular accrual periods.
+     * This property allows the stub rate to be specified, either as a known amount or a rate.
+     * If this property is not present, then the rate derived from the {@code rate} property applies during the stub.
+     * If this property is present and there is no initial stub, it is ignored.
+     * @param futureValueNotional  the new value
+     * @return this, for chaining, not null
+     */
+    public Builder futureValueNotional(FutureValueNotional futureValueNotional) {
+      this.futureValueNotional = futureValueNotional;
+      return this;
+    }
+
     //-----------------------------------------------------------------------
     @Override
     public String toString() {
-      StringBuilder buf = new StringBuilder(160);
+      StringBuilder buf = new StringBuilder(192);
       buf.append("FixedRateCalculation.Builder{");
       buf.append("dayCount").append('=').append(JodaBeanUtils.toString(dayCount)).append(',').append(' ');
       buf.append("rate").append('=').append(JodaBeanUtils.toString(rate)).append(',').append(' ');
       buf.append("initialStub").append('=').append(JodaBeanUtils.toString(initialStub)).append(',').append(' ');
-      buf.append("finalStub").append('=').append(JodaBeanUtils.toString(finalStub));
+      buf.append("finalStub").append('=').append(JodaBeanUtils.toString(finalStub)).append(',').append(' ');
+      buf.append("futureValueNotional").append('=').append(JodaBeanUtils.toString(futureValueNotional));
       buf.append('}');
       return buf.toString();
     }
