+++ modules/product/src/main/java/com/opengamma/strata/product/swap/NotionalSchedule.java
@@ -110,14 +110,6 @@ public final class NotionalSchedule
    */
   @PropertyDefinition
   private final boolean finalExchange;
-  /**
-   * The future value notional.
-   * <p>
-   * This property is used when the fixed leg of a swap has a future value notional.
-   * This is typically used for Brazilian swaps.
-   */
-  @PropertyDefinition(get = "optional")
-  private final FutureValueNotional futureValueNotional;
 
   //-------------------------------------------------------------------------
   /**
@@ -362,8 +354,7 @@ public final class NotionalSchedule
       ValueSchedule amount,
       boolean initialExchange,
       boolean intermediateExchange,
-      boolean finalExchange,
-      FutureValueNotional futureValueNotional) {
+      boolean finalExchange) {
     JodaBeanUtils.notNull(currency, "currency");
     JodaBeanUtils.notNull(amount, "amount");
     this.currency = currency;
@@ -372,7 +363,6 @@ public final class NotionalSchedule
     this.initialExchange = initialExchange;
     this.intermediateExchange = intermediateExchange;
     this.finalExchange = finalExchange;
-    this.futureValueNotional = futureValueNotional;
     validate();
   }
 
@@ -468,18 +458,6 @@ public final class NotionalSchedule
 
   //-----------------------------------------------------------------------
   /**
-   * Gets the future value notional.
-   * <p>
-   * This property is used when the fixed leg of a swap has a future value notional.
-   * This is typically used for Brazilian swaps.
-   * @return the optional value of the property, not null
-   */
-  public Optional<FutureValueNotional> getFutureValueNotional() {
-    return Optional.ofNullable(futureValueNotional);
-  }
-
-  //-----------------------------------------------------------------------
-  /**
    * Returns a builder that allows this bean to be mutated.
    * @return the mutable builder, not null
    */
@@ -499,8 +477,7 @@ public final class NotionalSchedule
           JodaBeanUtils.equal(amount, other.amount) &&
           (initialExchange == other.initialExchange) &&
           (intermediateExchange == other.intermediateExchange) &&
-          (finalExchange == other.finalExchange) &&
-          JodaBeanUtils.equal(futureValueNotional, other.futureValueNotional);
+          (finalExchange == other.finalExchange);
     }
     return false;
   }
@@ -514,21 +491,19 @@ public final class NotionalSchedule
     hash = hash * 31 + JodaBeanUtils.hashCode(initialExchange);
     hash = hash * 31 + JodaBeanUtils.hashCode(intermediateExchange);
     hash = hash * 31 + JodaBeanUtils.hashCode(finalExchange);
-    hash = hash * 31 + JodaBeanUtils.hashCode(futureValueNotional);
     return hash;
   }
 
   @Override
   public String toString() {
-    StringBuilder buf = new StringBuilder(256);
+    StringBuilder buf = new StringBuilder(224);
     buf.append("NotionalSchedule{");
     buf.append("currency").append('=').append(currency).append(',').append(' ');
     buf.append("fxReset").append('=').append(fxReset).append(',').append(' ');
     buf.append("amount").append('=').append(amount).append(',').append(' ');
     buf.append("initialExchange").append('=').append(initialExchange).append(',').append(' ');
     buf.append("intermediateExchange").append('=').append(intermediateExchange).append(',').append(' ');
-    buf.append("finalExchange").append('=').append(finalExchange).append(',').append(' ');
-    buf.append("futureValueNotional").append('=').append(JodaBeanUtils.toString(futureValueNotional));
+    buf.append("finalExchange").append('=').append(JodaBeanUtils.toString(finalExchange));
     buf.append('}');
     return buf.toString();
   }
@@ -574,11 +549,6 @@ public final class NotionalSchedule
     private final MetaProperty<Boolean> finalExchange = DirectMetaProperty.ofImmutable(
         this, "finalExchange", NotionalSchedule.class, Boolean.TYPE);
     /**
-     * The meta-property for the {@code futureValueNotional} property.
-     */
-    private final MetaProperty<FutureValueNotional> futureValueNotional = DirectMetaProperty.ofImmutable(
-        this, "futureValueNotional", NotionalSchedule.class, FutureValueNotional.class);
-    /**
      * The meta-properties.
      */
     private final Map<String, MetaProperty<?>> metaPropertyMap$ = new DirectMetaPropertyMap(
@@ -588,8 +558,7 @@ public final class NotionalSchedule
         "amount",
         "initialExchange",
         "intermediateExchange",
-        "finalExchange",
-        "futureValueNotional");
+        "finalExchange");
 
     /**
      * Restricted constructor.
@@ -612,8 +581,6 @@ public final class NotionalSchedule
           return intermediateExchange;
         case -1048781383:  // finalExchange
           return finalExchange;
-        case -282775858:  // futureValueNotional
-          return futureValueNotional;
       }
       return super.metaPropertyGet(propertyName);
     }
@@ -682,14 +649,6 @@ public final class NotionalSchedule
       return finalExchange;
     }
 
-    /**
-     * The meta-property for the {@code futureValueNotional} property.
-     * @return the meta-property, not null
-     */
-    public MetaProperty<FutureValueNotional> futureValueNotional() {
-      return futureValueNotional;
-    }
-
     //-----------------------------------------------------------------------
     @Override
     protected Object propertyGet(Bean bean, String propertyName, boolean quiet) {
@@ -706,8 +665,6 @@ public final class NotionalSchedule
           return ((NotionalSchedule) bean).isIntermediateExchange();
         case -1048781383:  // finalExchange
           return ((NotionalSchedule) bean).isFinalExchange();
-        case -282775858:  // futureValueNotional
-          return ((NotionalSchedule) bean).futureValueNotional;
       }
       return super.propertyGet(bean, propertyName, quiet);
     }
@@ -735,7 +692,6 @@ public final class NotionalSchedule
     private boolean initialExchange;
     private boolean intermediateExchange;
     private boolean finalExchange;
-    private FutureValueNotional futureValueNotional;
 
     /**
      * Restricted constructor.
@@ -754,7 +710,6 @@ public final class NotionalSchedule
       this.initialExchange = beanToCopy.isInitialExchange();
       this.intermediateExchange = beanToCopy.isIntermediateExchange();
       this.finalExchange = beanToCopy.isFinalExchange();
-      this.futureValueNotional = beanToCopy.futureValueNotional;
     }
 
     //-----------------------------------------------------------------------
@@ -773,8 +728,6 @@ public final class NotionalSchedule
           return intermediateExchange;
         case -1048781383:  // finalExchange
           return finalExchange;
-        case -282775858:  // futureValueNotional
-          return futureValueNotional;
         default:
           throw new NoSuchElementException("Unknown property: " + propertyName);
       }
@@ -801,9 +754,6 @@ public final class NotionalSchedule
         case -1048781383:  // finalExchange
           this.finalExchange = (Boolean) newValue;
           break;
-        case -282775858:  // futureValueNotional
-          this.futureValueNotional = (FutureValueNotional) newValue;
-          break;
         default:
           throw new NoSuchElementException("Unknown property: " + propertyName);
       }
@@ -824,8 +774,7 @@ public final class NotionalSchedule
           amount,
           initialExchange,
           intermediateExchange,
-          finalExchange,
-          futureValueNotional);
+          finalExchange);
     }
 
     //-----------------------------------------------------------------------
@@ -922,31 +871,17 @@ public final class NotionalSchedule
       return this;
     }
 
-    /**
-     * Sets the future value notional.
-     * <p>
-     * This property is used when the fixed leg of a swap has a future value notional.
-     * This is typically used for Brazilian swaps.
-     * @param futureValueNotional  the new value
-     * @return this, for chaining, not null
-     */
-    public Builder futureValueNotional(FutureValueNotional futureValueNotional) {
-      this.futureValueNotional = futureValueNotional;
-      return this;
-    }
-
     //-----------------------------------------------------------------------
     @Override
     public String toString() {
-      StringBuilder buf = new StringBuilder(256);
+      StringBuilder buf = new StringBuilder(224);
       buf.append("NotionalSchedule.Builder{");
       buf.append("currency").append('=').append(JodaBeanUtils.toString(currency)).append(',').append(' ');
       buf.append("fxReset").append('=').append(JodaBeanUtils.toString(fxReset)).append(',').append(' ');
       buf.append("amount").append('=').append(JodaBeanUtils.toString(amount)).append(',').append(' ');
       buf.append("initialExchange").append('=').append(JodaBeanUtils.toString(initialExchange)).append(',').append(' ');
       buf.append("intermediateExchange").append('=').append(JodaBeanUtils.toString(intermediateExchange)).append(',').append(' ');
-      buf.append("finalExchange").append('=').append(JodaBeanUtils.toString(finalExchange)).append(',').append(' ');
-      buf.append("futureValueNotional").append('=').append(JodaBeanUtils.toString(futureValueNotional));
+      buf.append("finalExchange").append('=').append(JodaBeanUtils.toString(finalExchange));
       buf.append('}');
       return buf.toString();
     }
